Generate paths from rational numbers
As you step through Euclid's algorithm for the rational number, one of
the variables that you maintain regenerates the corresponding path
vector!  The only issue is that rational numbers identify _two_
different continued fractions: one that ends in a 1, and one that does
not.  That means we can't currently distinguish 3.5.1 and 3.6 — they
both are represented by the same rational.