Add fudge factor to continued fractions
It looks like we should avoid including the number 1 in the continued
fractions that we create.  Rational numbers don't _uniquely_ identify a
_single_ continued fraction.  Rather, each rational number represents
_two_ continued fractions — one of which ends in a 1, and the other
which does not.

To get around this, we're just going to add 2 to each path vector
element before turning it into a rational number, and subtract it back
when regenerating the path representation.